codex_agent_op_v2
tree node id = 1dd31bad2ab33c72f664ddbb63c65d0efeeaafef9d436bdc348e9016bf8a4a29
test exit code = 0
unsafe count (v2) = 125

+++ translated_rust/src/gzread.rs
@@ -221,7 +221,7 @@ fn gz_look(state: &mut crate::gzguts_h::gz_state) -> ::core::ffi::c_int {
         // Keep both zero-filled buffers local until the inflater and their
         // registry ownership are ready. This preserves C's allocation order
         // while avoiding a partially-published state across setup.
-        let Some(input) = crate::src::gzlib::gz_owned_buffer(state.want) else {
+        let Some(input) = crate::src::gzlib::gz_owned_buffer(state.want as usize) else {
             crate::src::gzlib::gz_error(
                 state,
                 crate::zlib_h::Z_MEM_ERROR,
@@ -229,7 +229,7 @@ fn gz_look(state: &mut crate::gzguts_h::gz_state) -> ::core::ffi::c_int {
             );
             return -1 as ::core::ffi::c_int;
         };
-        let Some(output) = crate::src::gzlib::gz_owned_buffer(state.want.wrapping_shl(1)) else {
+        let Some(output) = crate::src::gzlib::gz_owned_buffer(state.want.wrapping_shl(1) as usize) else {
             crate::src::gzlib::gz_error(
                 state,
                 crate::zlib_h::Z_MEM_ERROR,
